r325102 introduced the split between pointer size (number of bytes / bits of a pointer) and index size (number of bytes / bits that are canonically used for GEP index operands).
This patch fixes a bunch of locations around ValueTracking where the
index size was incorrectly used for pointer types.

There are no new or fixed test cases. This patch simply fixes a
conceptual error that I noticed while working on ValueTracking for
pointers (for which I will submit a separate patch).

Change-Id: Ic63c2db644b53dc036ea2d1f037dafea4ad10b02